Sets up new ap-specific heatmap for the intrusion-sets dashboard
