--- src/app/intrusion-set-dashboard/attack-patterns-heatmap/attack-patterns-heatmap.component.html
@@ -1,40 +1,0 @@
-<unf-heatmap #heatmap [data]="heatMapData" [options]="heatMapOptions"
-        (hover)="highlightAttackPattern($event, true)" (click)="highlightAttackPattern($event)"></unf-heatmap>
-
-<ng-template #apTooltipTemplate>
-    <div class="col-md-12 attack-pattern-tooltip">
-        <mat-card>
-            <mat-card-subtitle>
-                <div class="flex1"><label>{{attackPattern.name}}</label></div>
-                <div [hidden]="hoverTooltip">
-                    <button mat-button type="button" class="mat-primary" *ngIf="isAdminUser()">
-                        <a target="_blank" href="/#/stix/attack-patterns/{{attackPattern.id}}">Details</a>
-                    </button>
-                    <button mat-button type="button" (click)="hideAttackPatternTooltip(attackPattern)">Close</button>
-                </div>
-            </mat-card-subtitle>
-            <mat-card-content>
-                <div class="row" *ngIf="attackPattern.description">
-                    <div class="col-md-3"><label>Description</label></div>
-                    <div class="col-md-8 attack-pattern-desc" [innerHTML]="attackPattern.description"></div>
-                </div>
-                <div class="row" *ngIf="attackPattern.x_mitre_data_sources && attackPattern.x_mitre_data_sources.length">
-                    <div class="col-md-3"><label>Data Sources</label></div>
-                    <div class="col-md-9 attack-pattern-sources">{{attackPattern.x_mitre_data_sources.join(', ')}}</div>
-                </div>
-                <div class="row" *ngIf="attackPattern.x_mitre_platforms && attackPattern.x_mitre_platforms.length">
-                    <div class="col-md-3"><label>Platforms</label></div>
-                    <div class="col-md-9 attack-pattern-platforms">{{attackPattern.x_mitre_platforms.join(', ')}}</div>
-                </div>
-                <div class="row" *ngIf="attackPattern.phases && attackPattern.phases.length">
-                    <div class="col-md-3"><label>Phases</label></div>
-                    <div class="col-md-9 attack-pattern-phases">{{attackPattern.phases.join(', ') | capitalize}}</div>
-                </div>
-                <div class="row" *ngIf="attackPattern.intrusion_sets && attackPattern.intrusion_sets.length">
-                    <div class="col-md-3"><label>Intrusion Sets</label></div>
-                    <div class="col-md-9 attack-pattern-intrusion-sets">{{getIntrusionSetNames(attackPattern)}}</div>
-                </div>
-            </mat-card-content>
-        </mat-card>
-    </div>
-</ng-template>
