--- src/app/intrusion-set-dashboard/attack-patterns-heatmap/attack-patterns-heatmap.component.scss
@@ -1,62 +1,0 @@
-@import '../../../styles/_variables.scss';
-
-$white: #FFFFFF;
-
-:host ::ng-deep {
-    unf-heatmap div.heat-map {
-        height: 500px;
-        background: $white;
-    }
-}
-
-.attack-pattern-tooltip {
-
-    padding: 10px;
-
-    .mat-card {
-        padding: 12px;
-        background-color: #f0f0f0;
-    }
-
-    .mat-card-subtitle {
-        display: flex;
-        align-items: center;
-        margin-bottom: 0;
-
-        label {
-            color: $dark-text-default;
-        }
-    }
-
-    div.row {
-        margin-bottom: 4px;
-    }
-
-    .attack-pattern-desc {
-        max-height: 60px;
-        overflow: hidden;
-    }
-
-    .attack-pattern-desc:after {
-        content: "";
-        text-align: right;
-        position: absolute;
-        bottom: 0;
-        right: 0;
-        top: 40px;
-        width: 70%;
-        height: 20px;
-        background: linear-gradient(to right, rgba(240, 240, 240, 0), rgba(240, 240, 240, 1) 80%);
-    }
-
-    .attack-pattern-sources {
-        max-height: 40px;
-        overflow: hidden;
-    }
-
-    .attack-pattern-platforms {
-        max-height: 20px;
-        overflow: hidden;
-    }
-
-}
