+++ src/app/intrusion-set-dashboard/intrusion-set-dashboard.component.ts
@@ -7,6 +7,7 @@ import {
         HostListener,
         ChangeDetectorRef,
     } from '@angular/core';
+import { Observable } from 'rxjs/Observable';
 
 import { MatDialog, MatDialogRef, MatSnackBar } from '@angular/material';
 import * as Ps from 'perfect-scrollbar';
@@ -17,7 +18,11 @@ import { BaseStixService } from '../settings/base-stix.service';
 import { BaseComponentService } from '../components/base-service.component';
 import { topRightSlide } from '../global/animations/top-right-slide';
 import { GenericApi } from '../core/services/genericapi.service';
+import { HeatmapComponent } from '../global/components/heatmap/heatmap.component';
+import { AttackPatternCell } from '../global/components/heatmap/ap-heatmap.component';
+import { HeatColor, HeatMapOptions } from '../global/components/heatmap/heatmap.data';
 import { IntrusionSetHighlighterService } from './intrusion-set-highlighter.service';
+import { Dictionary } from '../models/json/dictionary';
 
 @Component({
     selector: 'intrusion-set-dashboard',
@@ -32,7 +37,32 @@ export class IntrusionSetDashboardComponent implements OnInit {
 
     public groupKillchain: any[];
     public killChainPhases: any[];
+    private previousKillChainPhases = [];
+    private attackPatterns: Dictionary<any> = {};
+
     public showHeatMap = true;
+    public heatMapData: Array<AttackPatternCell> = [];
+    private noColor: HeatColor = {bg: '#ccc', fg: 'black'};
+    public readonly heatMapOptions: HeatMapOptions = {
+        color: {
+            batchColors: [
+                {header: {bg: '#4db6ac', fg: 'black'}, body: {bg: 'white', fg: 'black'}},
+            ],
+            heatColors: {'false': this.noColor},
+            noColor: this.noColor,
+            showGradients: true,
+            maxGradients: 3,
+            defaultGradient: {bg: ['#999', 'black'], fg: 'white'}
+        },
+        text: {
+            showCellText: true,
+        },
+        zoom: {
+            hasMinimap: true,
+        },
+    };
+    public hoverTooltip = true;
+    @ViewChild(HeatmapComponent) private heatMapView: HeatmapComponent;
 
     public showToolbox = false;
     @ViewChild('toolboxBtn') private toolboxBtn: ElementRef;
@@ -56,24 +86,73 @@ export class IntrusionSetDashboardComponent implements OnInit {
         protected genericApi: GenericApi,
         protected snackBar: MatSnackBar,
         protected ref: ChangeDetectorRef,
+        private highlighter: IntrusionSetHighlighterService,
     ) {}
 
     public ngOnInit() {
-        const APsortObj = { 'stix.name': '1' };
         const APprojectObj = {
+            'stix.id': 1,
             'stix.name': 1,
+            'stix.description': 1,
             'stix.kill_chain_phases': 1,
-            'stix.id': 1
+            'extendedProperties.x_mitre_data_sources': 1,
+            'extendedProperties.x_mitre_platforms': 1,
         };
+        const APsortObj = { 'stix.name': '1' };
         const APfilter = encodeURI(`sort=${JSON.stringify(APsortObj)}&project=${JSON.stringify(APprojectObj)}`);
-        const initData$ = this.genericApi.get(`${Constance.ATTACK_PATTERN_URL}?${APfilter}`)
-            .finally(() => initData$.unsubscribe())
+        const initAttackPatterns$ = this.genericApi.get(`${Constance.ATTACK_PATTERN_URL}?${APfilter}`)
+
+        const intrusionsProperties = {
+            'stix.id': 1,
+            'stix.name': 1,
+        };
+        const intrusionsFilter = encodeURI(`project=${JSON.stringify(intrusionsProperties)}`);
+        const initIntrusions$ = this.genericApi.get(`${Constance.INTRUSION_SET_URL}?${intrusionsFilter}`);
+
+        const relFilter = {
+            'stix.relationship_type': 'uses',
+            'stix.source_ref': {'$regex': '^intrusion\-set\-\-'},
+            'stix.target_ref': {'$regex': '^attack\-pattern\-\-'},
+        };
+        const relQuery = encodeURI(`filter=${JSON.stringify(relFilter)}`);
+        const initRelationships$ = this.genericApi.get(`${Constance.RELATIONSHIPS_URL}?${relQuery}`)
+
+        const initData$ = Observable.forkJoin(initAttackPatterns$, initIntrusions$, initRelationships$)
+            .finally(() => initData$ && initData$.unsubscribe())
             .subscribe(
-                (attackPatterns) => {
+                ([attackPatterns, intrusionSets, relationships]) => {
+                    attackPatterns.forEach(ap => {
+                        if (ap && ap.attributes) {
+                            const pattern = this.attackPatterns[ap.attributes.id] = ap.attributes;
+                            pattern.phases = pattern.kill_chain_phases.map(phase => phase.phase_name);
+                            pattern.analytics = [];
+                            pattern.values = [];
+                        };
+                        return ap;
+                    });
+
+                    const intrusions = {};
+                    intrusionSets.forEach(intrusion => {
+                        if (intrusion && intrusion.attributes) {
+                            intrusions[intrusion.attributes.id] = intrusion.attributes;
+                        }
+                    })
+
+                    relationships.forEach(rel => {
+                        if (rel && rel.attributes && rel.attributes.source_ref && rel.attributes.target_ref) {
+                            const ap = this.attackPatterns[rel.attributes.target_ref];
+                            const is = intrusions[rel.attributes.source_ref];
+                            if (ap && is) {
+                                ap.analytics.push(is);
+                            }
+                        }
+                    });
+
                     this.groupKillchain = this.groupByKillchain(attackPatterns);
                     this.killChainPhases = this.groupKillchain;
+                    this.createAttackPatternHeatMap();
                 },
-                (err) => console.log(new Date().toISOString(), err),
+                (err) => console.log(new Date().toISOString(), err)
             );
     }
 
@@ -90,7 +169,7 @@ export class IntrusionSetDashboardComponent implements OnInit {
                         killChainAttackPatternGroup[phaseName] = attackPatternsProxies;
                     }
                     attackPatternsProxies.push({
-                        name: attackPattern.attributes.name,
+                        id: attackPattern.attributes.id,
                         back: '#FFFFFF'
                     });
                 });
@@ -134,6 +213,7 @@ export class IntrusionSetDashboardComponent implements OnInit {
         if (!selections || (selections.length === 0)) {
             this.intrusionSets = null;
             this.killChainPhases = this.groupKillchain;
+            this.createAttackPatternHeatMap();
             this.treeData = null;
         } else {
             this.treeSpinner = true;
@@ -150,10 +230,48 @@ export class IntrusionSetDashboardComponent implements OnInit {
                         this.killChainPhases = data.killChainPhases;
                         this.totalAttackPatterns = data.totalAttackPatterns;
                         this.treeData = this.buildTreeData();
+                        this.createAttackPatternHeatMap();
                     },
                     (err) => console.log(new Date().toISOString(), err),
                 );
         }
+    }
+
+    /**
+     * @description Create a heatmap chart of all the tactics. This looks like a version of the carousel, but shrunken
+     *              in order to fit within the viewport.
+     */
+    private createAttackPatternHeatMap() {
+        let data = {};
+        this.killChainPhases.forEach(phase => {
+            let index = 0;
+            if (phase && phase.name && phase.attack_patterns) {
+                phase.attack_patterns.forEach(attackPattern => {
+                    if (attackPattern.id) {
+                        let ap = data[attackPattern.id];
+                        if (!ap) {
+                            ap = Object.assign({}, this.attackPatterns[attackPattern.id] || {});
+                            data[ap.id] = ap;
+                        }
+                        if (attackPattern.intrusion_sets && attackPattern.intrusion_sets.length) {
+                            ap.values = attackPattern.intrusion_sets
+                                .sort((a, b) => a.name.toLowerCase().localeCompare(b.name.toLowerCase()))
+                        } else {
+                            ap.values = [];
+                        }
+                    }
+                });
+            }
+        });
+        this.heatMapData = Object.values(data);
+    }
+
+    public highlightAttackPattern(selection: any) {
+        let ap = null;
+        if (selection && selection.row) {
+            ap = Object.values(this.attackPatterns).find(ptn => ptn.name === selection.row.title) || null;
+        }
+        this.highlighter.highlightIntrusionSets(ap);
     }
 
     /**
